Decouple login-test and issuer-frame-test from explicit soda import, as they can now pass in an extensions object as part of acceptance-test-utils.createClient().
